Implementation classes and new exception
I did the implementation classes for person and wallet. I also added a new exception for wallet. Lastly I fixed the repositories classes that had the incorrect type in the requirements of the jpa.

+++ src/main/java/com/technical/springboot/matepore/Springboot_Technical_Interview/repositories/WalletRepository.java
@@ -3,6 +3,6 @@ package com.technical.springboot.matepore.Springboot_Technical_Interview.reposit
 import com.technical.springboot.matepore.Springboot_Technical_Interview.entities.Wallet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WalletRepository extends JpaRepository<Wallet, Integer> {
+public interface WalletRepository extends JpaRepository<Wallet, Long> {
 
 }
